Change expand_lines() to take a single string argument
Change the method Change.expand_lines() to take its template argument
as a single string rather than requiring it to already be broken into
individual lines.  This means that the templates can be defined as
strings, which is simpler and less error-prone (especially for people
who want to modify the templates themselves).