Don't fail if "git config" does not support the --local option.
Matthieu Moy points out that older versions of git do not support the
--local option.  Try to work around its possible absence: if the call
to "git config --local" fails, then try again without "--local".  If
the second command indicates that the section is empty, then the local
version of the section is definitely empty, so delete it.

This might result in not deleting an empty local section (e.g., if
there is a non-empty global section with the same name), but that's
only a cosmetic problem.