make the Reply-To field configurable
This introduces multimailhook.{replyTo,replyToCommit,replyToRefchange},
that can be used for multiple purpose:

- disabling the notification

- setting the Reply-To to a particular email address (e.g. send the list
  to <project>-commits@list.server.com, but set the Reply-To to
  <project>-devel@list.server.com)

- changing the default strategy (pusher for refchange, author for
  commit).

We use the same variable names for all these scenarios, with the magic
values "author", "pusher" and "none". It is safe because the email
address has to contain a "@" and avoids introducing too many variables
for the same purpose.