Split up HEADER_TEMPLATES.
Split *_HEADER_TEMPLATE into *_HEADER_TEMPLATE and *_INTRO_TEMPLATE,
where the former contains only the email header lines and the latter
contains the text forming the beginning of the email body.  This (1)
will make it easier to handle the email headers in a more RFC
2822-compliant manner, and (2) is a better division for users who
might want to customize one or the other.