Properly handle header lines that are too long.
If a header line is too long, email.header.Header.encode() splits it
into multiple lines with proper line continuation characters.  Since
later code expects to receive lines one at a time, such lines need to
be re-split and yielded one by one.  The easiest way to do that is to
move the encoding functionality into Change.expand_header_lines(),
where (1) we can split the lines into name and value *before*
expansion and (2) we can emit as many lines as we want.  Also tell the
Header instance the name of the header so that it can account for that
length when splitting the first line.