Introduce multimailhook.useRepoPath to show path instead of repo name
Although the common practice is that the repository basename gives a
meaningful name, it can happen also to have directories organized like:

project1/git/
project2/git/
project3/git/

In this case, the default shortname will be [git] for all repositories.
This new option can be set in a users's ~/.gitconfig to have all
notification emails contain the full path to the repositories.