Add a simple test that the example post-receive hook script works.
Now that we can configure git-multimail to use a fake sendmail
program, it is possible to test the example post-receive script.  So
add a simple test that the script runs.

Extensive testing would be overkill, since it shares most of its code
with git_multimail.py.  But this one test should prevent silly
mistakes like missing imports or errors using the git_multimail API.