t/test-emails: set $TESTREPO from the output of create-test-repo
create-test-repo prints the $GIT_DIR of the repository that it just
created.  Use that output to set $TESTREPO in test-emails.  Aside from
removing the need to specify the name in two places, this avoids
letting the absolute path leak into the output of test-emails, where
it created a gratuitous variability in the output.