Pass a Change argument to Environment.get_fromaddr().
When Environment.get_fromaddr() is being called for a specific change,
pass it the Change instance for which it is being called.  This makes
it possible for users to override get_fromaddr() to return the
pusher/committer.

This change requires "fromaddr" to be taken out of COMPUTED_KEYS
because it has to be passed an argument, which means that it can no
longer be part of Environment.get_values().  So call get_fromaddr()
directly from Change._compute_values(), and adjust test-env
accordingly.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>