Don't use git's "-c" option if it is unavailable.
git's global "-c" option was not added until Git 1.7.2, whereas we
claim to support versions back to 1.7.1.  Therefore, check explicitly
whether the installed version of git supports the "-c" option, and if
not then omit it.

Currently we use "-c" only to set i18n.logoutputencoding=UTF-8.  The
omission of this setting is not necessarily a problem (for example if
the user hasn't overridden the default, which is UTF-8 anyway);
therefore it is reasonable to degrade to not setting this option.

Reported-by: GitHub user cliles-capario