Add a get_ref_ignore_list() function to allow refs to be ignored
Some deployments may have refs for which they don't want emails sent out
when they are updated.  However, if commits pushed to these refs are later
pushed to another ref, we do want emails for those commits at that time.
This means that these refs should be ignored both in terms of which
changes we should send out emails for, and when computing what commits are
considered new to the repository.

We set the default for this in the base Environment to ["refs/notes/"].
A default of [] would make sense if git_multimail had a specialized
formatting for commits made by git notes, but it currently doesn't.  In
particular, the commit messages of git notes are essentially meaningless
and "filenames" in git notes commits are merely an implementational
detail.  As such, sending emails for these commits, with our current
format, is suboptimal.

Notes are not the only reason for an exclusion, though.  For example,
Gerrit uses refs/changes/ to store commits that have not yet been vetted.
Users think of commits stored under refs/changes/ as being new to the
repository only when they are finally merged into a branch under
refs/heads/.  (There are a variety of reasons for this: (a) these commits
haven't undergone code review and thus are viewed as not yet having been
accepted, (b) Gerrit has a separate mechanism to handle things pushed to
refs/changes/ (emailing just a few relevant folks with a message along
the lines of "You have a Change Review that you need to look at; look at
this nifty web page to add your comments."), (c) only a small subset of
users are interested in any particular refs/changes/ push, whereas
hundreds of developers want to see what enters refs/heads/, (d), gerrit's
standard ref-updated hook, which is similar to git's update or
post-receive hooks, is not invoked for these refs.  It all adds up to
reasons to ignore this particular ref namespace.)  Another example is
Gerrit's refs/cache-automerge/, which is essentially just an
implementational detail that we clearly don't want to spam users about.

Signed-off-by: Elijah Newren <newren@palantir.com>