RFC: Silence loud error messages going to the error log
This is the wrong way to do this.  For plain git installations, or
gitolite installations, these messages are useful feedback sent to the
user when they do a push.  For stash, these messages are simply thrown
away.  For gerrit, they spam the error log.  There's no way I know of to
change stash or gerrit to send these to the user, and as they aren't
warnings or errors but are just informational, it's really annoying to
have these spam the error log.  Having a more reasonable way to turn
these messages off would be nice.

Signed-off-by: Elijah Newren <newren@palantir.com>