GerritEnvironment: Commits can enter repositories multiple ways
Since commits can both be directly 'pushed' to the repository, or first
be created as a patchset and then 'submitted' from the web UI (with the
latter far more common), claiming that all gerrit commits were 'pushed'
will confuse users.  Make an "update_method" value available for
templates so that emails can use the correct verb.

This would also apply to Stash, if it provided a way to differentiate
in its (Async)?PostReceiveRepositoryHook classes between being
triggered by a direct push and by the merge of a pull request.

Signed-off-by: Elijah Newren <newren@palantir.com>