Atlassian Stash plugin: Initial version
This adds a plugin for Atlassian Stash that will run git_multimail as an
AsyncPostReceiveRepositoryHook (similar to a standard post-receive hook,
but also fires for merges of pull-requests; further, via Stash's Java API
such a hook has access to important data like user and repo name).

This plugin currently provides a single textbox entry of configuration at
the time that a repository owner enables the plugin, in which the
repository owner can provide a comma-separated list of email addresses
that they want the emails sent to.  Other multimail configuration options
are unexposed through the web UI, though they can be configured on a
repository-by-repository basis on the server if an admin sets the
appropriate multimailhook.* configuration setting in the git config file.

Signed-off-by: Elijah Newren <newren@palantir.com>