GerritEnvironment, StashEnvironment: Make From=pusher by default
I see no value in getting emails from
   gerrit@<company-machine-no-one-needs-to-know-about>
or
   stash@<company-machine-no-one-needs-to-know-about>.localdomain
The default user.email setting is often far less than perfect, and having
to set multimailhook.from for every repository is not ideal either.

A default that makes more sense to me is that the email comes from the
person who pushed the change; for StashEnvironment we will always have
that information available and in GerritEnvironment we often will.

Signed-off-by: Elijah Newren <newren@palantir.com>