If a mailing list setting is missing, just don't send the mail.
Instead of being really insistent that all mailing list values are
configured, just treat unset values as "don't send this kind of
email".  In any case, the user will see a warning that no emails were
sent, so an accidental misconfiguration shouldn't go unnoticed for
long.

However, if no email recipients have been configured at all, abort, as
this is a situation that is of no use to anybody.