Emit single pushed commit as just a single email
No need to send both a refchanged email and commit email.

In the sadly-all-too-frequent usecase of people pushing only one of their
commits at a time to a repository, users feel the reference change summary
emails are noise rather than important signal.  This is because there is a
reference change summary email for each new commit, and all these
summaries do is point out that there is one new commit -- something that
is obvious from the existence of the individual revision email that is
also sent.  In such cases, users prefer there to be no push summary email.

Some examples where this is particularly important:

    Sadly, gerrit's review mechanism essentially enforces a workflow where
    every commit is submitted individually (things can be directly pushed
    if permissions are set up to allow it, but those pushes bypass the
    code review mechanisms).  Users can push several commits at once for
    review, but each commit is reviewed and submitted individually, and it
    is upon submit that gerrit merges the commit onto the real branch.
    This serialization means that gerrit's equivalent to the post-receive
    hook (named ref-updated) gets fired off for just a single new commit
    (well, two commits if you count the implicit merge commit created by
    gerrit, but users find emails about the implicit gerrit merge commits
    to be noise as well).

    Lots of people haven't weaned themselves from a cvs- or svn-like
    workflow may also push each commit individually.  For people who
    predominantly only push a single new commit, the push summary emails
    are just noise.

However, even in repositories where 98% of uses are one of the above,
there are still other folks that might sometimes do large merges between
branches, or push several commits at once, or potentially force update
some branches.  In all of these cases, the reference summary emails are
very useful and we still want them to be sent, so we do not want to just
turn off the reference summary emails on a repo-by-repo basis.  We instead
want to have them turned off just for the particular pushes for which
users deem them to be noise.

Eventually, the email to be sent will be a combined commit. Currently, we
just send the revision email and skip the summary email.

Helped-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Elijah Newren <newren@palantir.com>