squash! GerritEnvironment: Make git_multimail callable as Gerrit ref-updated hook
The stuff gerrit stores under refs/meta/config is also an implementational
detail of gerrit, which stores access control information.  Users aren't
familiar with the file format, and it is displayed in the Gerrit UI in a
different format.  Users probably don't want to be emailed when access
controls for a project are modified (that's admin-privileged information,
and they have a different UI for viewing it), so just filter this stuff
out.