post-receive: catch ConfigurationException
The catch is already done for you if you use git_multimail.py directly.
But if one whishes to use post-receive as a template, it is probably a
good idea to catch the exception here too.

Leaving the exception uncaught could make sense too (to give more
information to the developer), but it's easier to notice the try/catch
and remove it than to guess that one should be there when it's not in the
template.

Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>