clean up code that calculates new and discarded commits, 1 of 7
convert `Push._compute_other_ref_sha1s()` into a property

This is the first in a series of commits that will enable reuse of the
`git rev-list` arguments that are used to determine the commits that
were added to or discarded from the repository.  Eventually these
arguments will be reused to include a graph of new/discarded commits
in the summary email.

Signed-off-by: Richard Hansen <rhansen@rhansen.org>