IMAP: authenticate when a login or password is provided
The previous code was calling stmp.login() in ssl mode. ssl is not the
only mode which requires authentication (we're going to add tls in the
next patch).

Actually, I'm not even sure the requirement to have a login and password
is a good thing (I think it's possible to have a certificate-based
authentication instead of a password-based one), but I don't have a way
to test it.