quote variable expansions that are subject to field splitting
Not all of these need to be quoted, but it's good to be in the habit
anyway.  Otherwise, a typo could cause the expansion to be the empty
string, which, if unquoted, causes it to act as if the argument wasn't
there.  Having it expand to an empty argument is better than no
argument because an empty argument often generates a noticeable error.

Signed-off-by: Richard Hansen <rhansen@rhansen.org>