fix check of rstcheck's exit status
By piping the output of rstcheck to tee, rstcheck's exit status
(return value) is lost so it can't be used to check for error.
Instead of piping to tee, simply redirect its output to a file.  If
that file ends up non-empty, cat its contents so that the user can see
what went wrong.

With this change, the user won't see any of rstcheck's output until
rstcheck has exited.  Fortunately, rstcheck is fast enough that the
user shouldn't notice.  If the delay does become a problem, we'll need
to use some fairly complex command substitution and redirection magic
to capture rstcheck's return value while simultaneously watching for
output.

Signed-off-by: Richard Hansen <rhansen@rhansen.org>