use common helper functions instead of 'echo'
Use 'log', 'error', or 'fatal' as appropriate instead of 'echo' when
logging a message for the user.  Use 'pecho' for all other uses of
'echo'.

(According to the POSIX specification, 'echo' may or may not process
backslashes and may or may not treat the first argument differently if
it begins with '-'.  'pecho' is a printf-based alternative that
behaves consistently on all platforms.)

Signed-off-by: Richard Hansen <rhansen@rhansen.org>