rstcheck: display output, and do check for exit status
Ironically, I'm the one who said that the exit status was irrelevant not
long ago, but it turns out that sharness partly breaks rstcheck for me:
it sets $HOME and my installation of rstcheck requires $HOME to be
properly set. Without this patch, I was getting a failure without any
explanation. Now, I'm getting the python stack trace. And I've repaired
my installation of rstcheck ;-).

Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>