Implement git-multimail -c, like git -c var=val
This implements the feature request in issue #102.

This allows passing configuration settings to git-multimail on the command
line, implemented by setting git's GIT_CONFIG_PARAMETERS environment
variable so that when we call 'git config --get ...', it'll return the
appropriate values.

This also simplifies the test scripts since we can avoid using additional
pre- and post- commands to set and unset configuration values for a single
run.

Signed-off-by: Elijah Newren <newren@gmail.com>