.travis.yml: don't install rstcheck for unsupported versions
The important is that the RST files are checked at least once, but we
don't really care about the Python version. Python 3.2 does not seem
supported at all, Python 2.6 is not officially supported by seems to work
with version 1.0. Stop bothering about these and skip the rstcheck tests
for Python 2.6 and 3.2.

Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>