log_*: adapt behavior to encoding and Python version
Sharness sets LC_ALL=C by default, which prevents Python 3 from outputing
non-ascii text, which we sometimes do. When the default encoding is
ascii (which is enforced by the default setting of Sharness), we'll fall
back to UTF-8, which is a saner default than outputing a badly encoded
string as Python 3 was doing by default.

We still use the default encoding by default, in case someone tries to
use git_multimail.py with a non-ascii and non-UTF-8 encoding (although
the user would be doing this at his own risk ...).

Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>