doc/gitolite: mention gitolite's UNSAFE_PATT
Fixes #119.

The documentation is kept short and is purposely not sufficient for the
user to set UNSAFE_PATT, as I do not want people to blindly apply an
unsafe reciepe. Instead, add a link to the official documentation of
UNSAFE_PATT which explains the safety implications better than we would
do here.

Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>