t/email-content, create-test-repo: don't cd to test dir
The Sharness way to run tests is to let sharness.sh create a trash
directory and cd to it. We were running inside test-repo.git which lived
outside this trash directory.

Start doing things the Sharness way, and let Sharness decide whether to
delete the trash directory at the end: we don't need the cleanup test
anymore. One positive side effect is that the trash directory is not
deleted anymore when a test fails.

Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>