simplify the logic of get_recipients
We previously had Config.get_recipients dealing with the
"comma-separated" logic, and
ConfigRecipientsEnvironmentMixin._get_recipients() to find the first
non-empty variable in a list.

Config should be only about calling "git config", and should not start
the git-multimail specific processing. Get the "comma-separated" logic
out of the Config class and merge it with
ConfigRecipientsEnvironmentMixin._get_recipients().

Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>