make HTML output better for text-based browser
My use-case is to be able to open it in a mailer like Gnus with w3m in
Emacs. Before this commit, the patch was displayed with one blank line
between every lines: the individual <pre> used for each line were
creating this extra-space.

Simplify the output by having one <pre> block for the whole code.
<div> is not allowed inside <pre>, hence use <span> instead to color
individual lines.

The result still displays fine in Firefox, but also looks good in my
mailer now.

Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>