remove unused variable from_addr
As found by pyflakes, variable from_addr was set but never used.
Actually, the precedence from{Refchange,commit} is dealt with in
ConfigOptionsEnvironmentMixin.get_fromaddr(), which is the one running
config.get('from'). So, this line was indeed useless.

The line was introduced in 8a97748 (Apply address processing to From:
field, 2015-06-19) where it was already unused.

Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>