SMTPMailer: catch exception on quit()
We're running smtp.quit() after an exception was raised. Since there was
an error earlier, it's likely that the connection is already in a broken
state and smtp.quit() may fail. Not catching the exception means
throwing a full backtrace to the user, which may hide the origin of the
real problem a few lines above.

Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>