shutdown the SMTP mailer before exiting
In Python 3, the following error was happening:

Exception ignored in: <bound method SMTPMailer.__del__ of <__main__.SMTPMailer object at 0x7f7a2172c1d0>>
Traceback (most recent call last):
  File "../../git-multimail/git_multimail.py", line 2002, in __del__
    self.smtp.quit()
  File "/usr/lib/python3.4/smtplib.py", line 868, in quit
    res = self.docmd("quit")
  File "/usr/lib/python3.4/smtplib.py", line 404, in docmd
    return self.getreply()
  File "/usr/lib/python3.4/smtplib.py", line 369, in getreply
    line = self.file.readline(_MAXLINE + 1)
ValueError: readline of closed file

It wasn't very serious since the expection happened within the
destructor hence was ignored, but that still showed a scary message to
the user.

The error happened at process exit time. Probably the SMTPMailer's
destructor was called after the socket or file was closed. There doesn't
seem to be a clean way to detect this.

Instead of letting the process termination call the destructor, call it
ourselves. git_multimail.py is the one which created the SMTP objet, so
it is legitimate to destroy it ourselves.

Fixes #141.

Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>