t/check-sign-off: work around git log --no-merge's behavior
In shallow clones, if the last commit in history is a merge, it has two
parents but none of them is present in the repository. In this case,
"git log --no-merge" shows the commit, and if this merge happens to lack
a sign-off (as are merge commits created by GitHub's "merge" button), we
got a false positive.

Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>