new multimailhook.commitBrowseURL variable
It was already possible to include a link to a repostory browser using
multimailhook.commitEmailFormat=html, multimailhook.htmlInIntro=true,
and overloading the REVISION_INTRO_TEMPLATE. This is flexible, but far
more complex than it needs to be for the user.

This commit adds multimailhook.commitBrowseURL that can be set to a URL,
possibly customized using format directives and a link will
automatically be added to each commit email.

Fixes #139.

Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>