logging infrastructure
This will allow more customization on the output of git_multimail.py
(just the normal info, or more output for debugging), and sending some
output to a log file (very useful to debug hard-to-reproduce issues
where one wants to know a posteriori whether git_multimail.py failed
to send a mail or the mail was lost later for example).

Tests are still very weak since the infrastructure is not really used
for now.

Fixes #155.
Fixes #148.

Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>