Make sure that multimailhook.from option has the highest priority.
Than any other way to compute "From:" address of the email. In other words,
if that option is set in a standard git config for a repository, a
notification for that repository with go out with that value in "From:"
field in email, period. This includes accessing that repository by different
means (corresponding to git-multimail "environments"). For example, a
repository accessed via normal git daemon and then via Gerrit, Stash,
or whatever will behave consistently.

The motivation for this change is that in many cases, repositories used
with git-multimail are forks of open-source projects, containing work
of hundreds people. If an organization X sets up a mirror/fork of such
a repository, it should avoid by all means trying to generate emails
impersonating completely unrelated persons who contributing to an
upstream project (for example, by using commit author information in
the "From:" field of a notification emails). In other words, there
should be a global-override style setting which, being set once, which
preclude any possibility of impersonation (which can be considered
unethitical behavior) once and for all, regardless of any other
variable factors. With this patch, multimail.from is exactly such
setting.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>