avoid dynamic choice of mixins in choose_environment
Add CLIRecipientsEnvironmentMixin to deal with the --recipients options
with the same mixin pattern as the others.

The eventual goal is to get a cleaner separation between building the
EffectiveEnvironment class (which should depend on the environment, but
not on how git_multimail.py is called) and instanciating it.

This will allow cleanly creating the classes while loading
git_multimail.py is loaded and before any user code is executed.

Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>