move environment checks to a separate check() method
Raising errors in the constructor causes many trouble: one needs to have
all the information available in the constructor, while the full class
is not built. For example, scancommitforcc was not normally available in
the constructor or ConfigEnvironmentMixin but it was needed for the
check hence its value was duplicated and it had to be special-cased in
choose_environment(). The config mixin is not even built, hence we can't
call any method accessing self.config in the constructor (even after
super(...).__init__(...) has completed).

Move the check to a separate method, and call it after the constructor
has finished building the class.

Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>