Fix typo in class name
Classes corresponding to different environments are built dynamically
using type(name, parents). The name of the class was built with a
typo. This was harmless in most cases, since we used the class itself
but rarely its name. As a consequence, the typo went undetected by the
testsuite.

Commit message and CHANGES added by Matthieu Moy
<git@matthieu-moy.fr>.

Signed-off-by: David Lazar <lazard@csail.mit.edu>
Signed-off-by: Matthieu Moy <git@matthieu-moy.fr>