fix migrate-mailhook-config
The Config.get_recipients() function disappeared in 5706012 (simplify
the logic of get_recipients, 2015-08-25), but is still used in
migrate-mailhook-config (for which we don't have automated testing ...).

Fix this by inlining it's old content (except for an emptyness test
which isn't relevant because we're inside an 'if' that makes sure the
list isn't empty).

An alternative would be to re-introduce the function, but inlining makes
sure we won't break migrate-mailhook-config again the same way.

Fixes #199.

Signed-off-by: Matthieu Moy <git@matthieu-moy.fr>