Introduce excludeMergeRevisions config variable
We switched to using git-multimail for sending mail to
git-commits-head@vger.kernel.org, but in the process found out that
there was no way to filter out merge commits, which kernel developers
don't want to see on that mailing list. This change introduces another
config variable called excludeMergeRevisions, which achieves the
functionality we needed.

Additionally, this adds the "parents" info to the values that can be
expanded in the templates (off by default).

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>