CI: checkout with a depth of 5, not the default 1.
By default, GitHub actions checks out only the latest version in a shallow
clone. This doesn't work for us, since our testsuite runs git-multimail on its
own history and HEAD^ needs to exists for 00-sanity.t to work. A depth of 2
would be sufficient in a linear history, but in case the latest commit is a
merge commit, don't rely on Git cloning exactly the 2 commits HEAD and HEAD^,
and take a few more commits.

Signed-off-by: Matthieu Moy <git@matthieu-moy.fr>