Add Gitea environment to improve emails sent when using it
Because Gitea and its fork Forgejo run the hook as the user they are
themselves running under, by default no email address of the commit or
push author was used. Change this by creating an environment extracting
this information from Gitea-provided environment variables and use it
automatically if they are defined.